+++ baybe/constraints/__init__.py
@@ -16,6 +16,7 @@ from baybe.constraints.discrete import (
     DiscreteNoLabelDuplicatesConstraint,
     DiscretePermutationInvarianceConstraint,
     DiscreteProductConstraint,
+    DiscreteRepetitionConstraint,
     DiscreteSelectionConstraint,
     DiscreteSumConstraint,
 )
@@ -34,11 +35,12 @@ __all__ = [
     "DiscreteCustomConstraint",
     "DiscreteDependenciesConstraint",
     "DiscreteExcludeConstraint",
-    "DiscreteSelectionConstraint",
     "DiscreteLinkedParametersConstraint",
     "DiscreteNoLabelDuplicatesConstraint",
     "DiscretePermutationInvarianceConstraint",
     "DiscreteProductConstraint",
+    "DiscreteRepetitionConstraint",
+    "DiscreteSelectionConstraint",
     "DiscreteSumConstraint",
     # --- Other --- #
     "validate_constraints",